LinkedListRange: track last added node
When a new node is added to the list, it is maintained as the last added node.

Inturn when removing a node, if it happens to be the last added node, then the
last added node is set to NULL.

Now ll_print also prints the content of Start and LastAdded nodes at the end.