LinkedListRange: ll_add_sorted_startfrom_lastadded
The original ll_add_sorted renamed to ll_add_sorted_startfrom_start, because
that logics starts checking from the start i.e begining of the list, to decide
a suitable place to add the new node.

Have added ll_add_sorted_startfrom_lastadded which starts checking from the
last added node, to decide a suitable place to add the new node. However if
llLastAdded is NULL, then it falls back to ll_add_sorted_startfrom_start.

For now,
the original name ll_add_sorted is #defined to ll_add_sorted_startfrom_start.

For the current lost packets range based linked list usage, based on how those
lost packets are identified and added, the startfrom_lastadded should provide
the right and most efficient way of adding to the linked list.

HOWEVER TO ACHIEVE A MORE GENERIC Double Linked List usage:

However at a later date, I may provide a more optimised implementation, which
initially starts looking around the lastAdded node, and if it doesn't find a
set of nodes, in-between which this new node has to be added,

then it will look around the end node (logic to maintain reference to end node
to be added yet) to see if it can find a pair of nodes in-between which the new
node has to be added,

if still not found, then start looking from the starting node.

TODO:LATER2: To help track the end node in the list,

I have to check in _ll_add, just before returning, if the newly added node's
next reference is NULL or not. If NULL, then the new added node is the end
node.

And in _ll_delete, I have to check if the node being deleted is the end node,
if so update the end node to point to the node previous to this node being
deleted.