Client logic uses ll_add_sorted_startfrom_lastadded
The way in which packets are sent from server as part of mcast test/transfer,
And inturn the way the client identifies lost packets, which currently
technically ignores any out of order packets (which currently based on
practical experiments is found to be almost 0/nil).

The logic will always generate lost-packets ranges which keep increasing
without overlapping, so ll_add_sorted_startfrom_lastadded is the best logic to
use to add to the linked list of lost packets ranges.