LinkedListRange: keep track of the end node
i.e keep track of the last / end node in the list.

When ever a new node is being added, if it is at the end of the list, then
update end node reference.

When ever a node is being deleted, if it was the end node already, then set end
node reference to the deleted nodes prev reference.

The ll_init takes care of setting end reference to NULL.

ll_print takes care of printing the contents of the end node.