LinkedListRange: TotalFromRanges - s1
Added initial logic to track the total/amount of values stored in the list, by
using ranges.

Initially iTotalFromRanges is initialised to 0.

As nodes are added, the amount/values/numbers stored as range in those nodes
are added to iTotalFromRanges.

i.e
if a Node with range 1-1 is added, iTotalFromRanges is incremented by 1.
if a Node with range 1-5 is added, iTotalFromRanges is incremented by 5.

ll_print now prints the value stored in iTotalFromRanges.

TODO: Have to update iTotalFromRanges suitably as things/values are deleted
from the list using ll_delete. In this case we require to keep decrementing.