CurRemaining LostPkts info, More ranges per query
Have increased the buffer alloted for communicating to server in Unicast
Recovery logic to help with following

a) The number of lost packets ranges that is communicated is increased from 10
to 20.

Inturn the server side guess_numofattempts logic also updated wrt same.

b) Also now the still remaining amount of LostPacket Ranges in the list as well
as the corresponding total amount of lost packets is now communicated to
server.

Inturn the server side logic updated to extract this info from the URAckSeqNum
response packets and print to console for now.