Optimise ll_delete - core & starthint_start
Restructuring linked list delete logic to provide optimised versions

a) The original ll_delete is now called ll_delete_core and inturn instead of
always starting from llStart, it now starts from the node explicitly given to
it.

b) added a new ll_delete_starthint_start, which calls delete_core with the node
to start checking from to llStart. This is same as the previous behaivour of
ll_delete. In turn ll_delete is defined to map to this.

TODO: Add a ll_delete_starthint_beforedel