LinkedListRange ll_delete_starthint_beforedel - p1
Now when ever a value is being deleted from the linked list, it will maintain a
reference to the node which comes before the node where the deleted value was.

In-turn added a new delete func called ll_delete_starthint_beforedel, which
asks the delete_core function to start searching for value to delete from the
llBeforeDel node in the list, provided it is not NULL.

TODO: If the user passes a value which would come in a node much before the
llBeforeDel node, then it would be missed. Make this func intelligent to detect
and start the search from beginning in such a situation.

NOTE: This function optimises deleting from the linked list, when the values
being deleted will be such that they keep increasing in their value.

Once the issue mentioned in TODO above is fixed, then it can be used in any
situation. And potentially if required ll_delete could be defined to map to
this.