Send TotalBlocks involved with MCastStop command
If the client doesnt recieve the last few adjacent blocks (how ever large or
small the number of lost blocks at the end may be), then in current logic it
wont know that it has lost those.

So inturn during Unicast recovery, when server uses URReqSeqNum packets /
commands to help the client recover lost packets, the client wont request for
those lost packets at the end as part of its URAckSeqNum response.

To help resolve this above issue, now the server sends the total blocks
involved in the current transfer, as a member of the MCastSTOP command packet.
It will be stored as a 4byte value starting from offset 8.

To achieve the same, now in file mode, the logic gets the file size and convert
it into number of blocks using dataSize as the individual block size.

It also checks to ensure that the total blocks involved doesnt exceed the
positive 2^31 bit space available for a signed 32bit integer. i.e a file or
test transfer of around 2TB (i.e 2e9*1024) can be transfered.