Client: Account for lost packets if any at the end
Now the logic keeps track of the largest seq number packet, that it has got.

During MCastSTOP, now it extracts the actual last packet that it should have
got during this transfer, from the info in the MCastSTOP packet.

In turn if required it updates the lost packets list.

As mentioned in the last commit, this takes care of a known corner case which
was there in the logic.

Now the info related to above is printed as part of mcast stop notification
message printed on the client side.