Fix oversights - LargeFiles, PktDataSize, int
*** In the Client code:

As I will be seeking beyond 2GB in a given file, I require to either enable
_FILE_OFFSET_BITS=64 or enable _LARGEFILE64_SOURCE and use lseek64.

Also when calculating the offset to seek, If the arguments to the right of the
assignment are all 32bit ints, then even if the left hand side is a 64bit int,
the calculation will generate a 32bit value and sign extend it, thus generating
a wrong offset value. So one is required to explicitly cast atleast one of the
elements on the rhs of assignment in the calculation to off64_t in this case.

Have done both of above.

Also previously when calculating the seek offsets for the blockIds, I was
wrongly using the packet size instead of the data size with in the packet. So
each block was getting written further away from its original position in the
file. Have fixed this.

*** In server side:

I had forgotten to make default test blocks value into int when assigning to
giTotalBlocksInvolved. This was leading to a struct pack failure in MCastSTOP
logic, have fixed this now.