check-image made less verbose and more intelligent
Now it prints the block id only if the expected value and got value wrt the
marker byte at the beginning of each dataSize block is different.

Parallely it also prints block id periodically, to indicate the current
progress.

It also prints the block id and delta, if there is a difference of more than 1
between two adjacent blocks