Srvr:LoadClients from saved/manual_created Context
NwRecover script now uses the new context module to load a list of clients it
should work with from a context file.

This context file

could be manually created by the user, if they know the list of clients one
should work with.

OR

it could be created by the program previously, potentially the list of
remaining clients, when the program in a previous run set aside after multiple
attempts to allow them to recover their lost packets/data.