Testing: Updates to few test related scripts
The Qemu virtual machines used for testing now have a addition harddisk which
is slightly more bigger.

Added some additional files which we may want to tranfer to the VMs, but by
default they are not transfered, as they are kept after the quit.

Also trying to get qemu-tftp-get-files.input if already using same to get the
files, may or may not go properly. Ideally it should be fine, but not sure if
there is a corner case in parallel file handling somewhere in the flow.