Testing: Updated message printed wrt TotalBlocks to check
Also added a new command to generate a 10GB file with holes except for the last
1024 block. This is similar to the 5GB file which I created previously.

With these holes in the file, one can get the effect of a large file, without
actually requiring that much space to be available on the filesystem. i.e in
filesystem only 1024 bytes will be used these 5GB and 10GB files.

NOTE: However the simpnwmon02 will generate a actual 5GB or 10GB file, so
remember to have space in the filesystem into which simpnwmon02 may write such
files.