Srvr:PI: show which new, which old and performance
Now that the user can give a context file to nw-recover logic.

There will be a client list independent of PI. So the PI logic now keeps track
as to which and all clients are known from beginning, which and all are new
clients and how either of them worked wrt PI interms of how many PIReq commands
came from the clients (new or known).