Srvr:mcast stop moved to new network modules
Created a new network module.

All network related logics will be consolidated into this module in the long
run.

For now have moved the mcast stop logic and related global constants into this
module.

This is to help with using the same logic also from unicast recovery related
script. The reason for which is mentioned in the TODO file.