Srvr:UR: Give less time to non-responding clients
If there are 35 clients, then if 50% of them are not responding for some reason
then previously it would have taken

18 Clients * 3 Minutes = 54 Minutes of wasted time

Now

with the wait time per client reduced to 1.5 minutes, it now wastes 27 minutes.

For more clients have to see how many parallel 35 client groups can be served
by a single server.