Srvr: SimLoss Random logic added
Previously the SimLoss logic was working on a fixed logic, which always losses
the same packets.

However now there is a variable bSimLossRandom which is by default set to true.

And inturn if this variable is true, then the packet losses will be random
interms of how often packets will be lost and how many packets can get lost at
a given instance.

NOTE: It is random within a given range for both aspects mentioned above.

If this variable is false, then the old SimLoss logic equivalent is what still
runs.