LinkedListRange ll_load partial implement & testing
ll_load has been partially implemented. It reads the file and extracts few of
the ranges and prints them to console.

As of now the exit condition is based on whether reading of file reached end of
file or not. If I change it to whether buffer is empty or not, it should load
more of the ranges.I