LinkedListRange Print all ranges during loading
As mentioned in previous commit, instead of checking if read reached end of
file or not, now I check if the buffer in memory which contains read data from
file has become empty due to no more file data as well as extraction of all the
ranges in it.

Also now I skip the newline at the end of each range, instead of letting strtol
do it for me.