LinkedListRange ll_free consistancy check cleanup
Now ll_free updates iNodeCnt and iTotalFromRange as it keeps freeing up
individual nodes. So at the end of ll_free both should be 0, else something is
wrong. Do cross verify this, when ever using, if things don't seem fine.

Also have taken care of setting llLastAdded and llBeforeDel to NULL, when they
get freed. Technically they can be set to NULL at the end in one shot, instead
of the current logic.

llEnd is set to NULL at the end, as it gets freed at the end either way.