Client: asked/forced to quit save LostPacketRanges
Added a signal handler for SIGINT, which calls ll_save to save the current list
of lost packet ranges to a predefined file (for now).

One can send SIGINT or press ctrl+c in console to trigger this.

TODO:

With this next I can add logic such that the program can resume from where it
had left/forced_to_quit potentially.

a simple resume ==>
For this at a minimum I have to pass a additional argument, which if passed can
be the file containing the saved lost packet ranges list. In turn by calling
ll_load and directly going into presence_info phase by bypassing the mcast
phase, I can implement a simple resume logic.

With this one can quit after mcast or in the middle of ucast recovery and still
be able to resume at a later stage, without any in-efficiencies or corner
cases. i.e the simple resume will still be the optimal way of resuming for such
cases.

Only in the case of quiting in the middle of mcast transfer, a simple resume
will lead to a inefficient recovery, because it will not recieve any additional
transfer from the mcast phase, even thou mcast transfer may still be on, when
the client resumed. However even here the logic will fully recover the contents
for such a client by using the ucast recovery mechanism.

a fully optimal resume ==>
However If I require to resume even mcast transfer phase, then I will have to
update the pktId and any additional required context of mcast recv logic. This
will also require additionaly saving the current phase, when the program was
forced to quit, so that the resume can be efficient.

In turn when resuming, using the current phase info saved, the logic can decide
whether to resume into mcast recv or into ucast recovery (potentially thro PI
phase or directly into ucast recv related phase).