FileUtils fixed a oversight wrt readline logic
If there are is no more data in the intermediate buffer, fresh data is read
from the file. But in that situation I had forgotten to reset the current
position pointer to the beginning of the buffer. Have fixed this now.

NOTE: With this that bug with ll_load not loading ranges towards the end of the
file has been fixed.

TODO: But wrt this particular case, have to think if I can use strtol's result
to identify issues if any with readline and flag a warning. i.e if this fix is
not applied, can ll_load identify issue with readline goof up, has to be
checked and implemented.