FileUtils readline eof/error not necessarily empty buf
There was a oversight in readline logic, where if the reading of file returns 0
or -1 (indicating eof or error), I was returning the same back to the caller.

If read returns eof, then we should return the buffer read till that point in
time. If there was no existing data in intermediate buffer, then the returned
buffer will also be empty and value returned will be 0, which even iPos will
match. However if there was some data in intermediate buffer when readline was
called, then it would have already got copied into the caller supplied buffer.
And as newline was not found in that data, will be the reason why the logic
would have been trying to read more data from file when the read returned 0, so
in this case the data copied into caller supplied buffer till now is what
should be returned and this inturn means the returned value will match iPos at
that time, in the current code flow. And next time when the caller calls
automatically 0 would be returned, as there will no longer be any data in
intermediate buffer and file also doesn't have any more data.

If read returned -1, then we could either

* send -1 OR

* send size of copied buffer till that point in time. Which would be either 0
  (if there was no data in intermediate buffer, when readline was called) or
  some size initially followed by 0 in the subsequent calls, the reasons for
  this is same as that explained wrt eof above.

As long as the caller handles the return value suitably, either of two options
above can be implemented wrt file read returning error situation.

Currently in both cases mentioned above i.e eof and error, we return iPos.

So in case of eof and error, one will get 0 as the return value from readline,
after any intermediate data as already been returned to the user once.