client save context func
Have moved the save context logic from signal handler into a seperate function,
which is inturn called from the signal handler.

With this the save context logic can be called from any location at anytime. To
help identify and keep it seperate from one another now, I pass a tag which is
put as suffix to the filename which contains these saved context.

This allows one to take a base path+filename as a argument and then generate
multiple save_context file as required with different suffixes.

NOTE: For now the save context only contains the list of lost packet ranges at
that time.