Client: AllInclusive Contxt initialStep, parseArgs
Added a structure on the client side, which will contain all the required
context across phases.

Currently it contains the data which is taken from the user over the command
line.

Inturn made the commandline arguments to be specified in a more verbose,
flexible and clear cut way. Added a new parse_args function to do the required
work.