Client: Cleanup and --runmodes
Added a new commandline --runmodes which tells which and all phases to run on
the client side.

	0 = Dont run any phases
	1 = Run MCast phase
	2 = Run UCast Presence Info phase
	4 = Run UCast Unicast Recovery phase
	6 = Run all UCast phases (PI+UR)
	7 = Run all phases (MCast + UCast)

In turn updated the main flow to satisfy this iRunModes.

Now the mcast specific/related context is saved only if mcast phase is run.

If --context context_file is specified and inturn loading of the context fails,
then the program exits, rather than the prev logic, where it will continue by
making mcast phase also to run.