Client parallel nw/prg instances = nwGroup support
To run multiple instances of the nw_test/data_transfer logic, I require to have
a mechanism for each instance of the server or client (or a group of clients on
different machines) program to run using a different set of client and server
ports.

So to help achieve this, I have added the concept of a NwGroup number. This
NwGroup id/number is used to determine a unique set of ports for each unique
NwGroup.

One requires to give the server and its associated list of clients the same
NwGroup number.

Have added this support to client side, next I require to add the same to
server side also.