Client: Added iMaxDataSeqNumGot to the snm context
The max seq num that has been got wrt the Data packets is now saved into the
snm context, previously it was local to the mcast recv.

THis is used to decide are there any packets lost at the end, when mstop is
got.

Now the iSeq and iPrevSeq is set at the beginning of the mcast recv logic,
based on the max data seq num seen till that time. This allows the mcast recv
to work irrespective of whether it is being called at the beginning of a
transfer or in the middle of the transfer or in the middle of the transfer
after it was stopped sometime before wrt the same transfer.

If the client is being called for the 1st time wrt a transfer then the max data
seq num got till date should be -1. However if the client is being resumed in
the middle of the mcast transfer, after it was previously stopped in the middle
of the same mcast transfer, then it should be the last seen max data seq num it
had got before it got stopped. This will allow any packets lost between the
interruption and resumption to be properly saved into the lostpacket ranges
list, and inturn recovered during the recovery phase.

Now if this aspect of snm context is saved while quiting and restored while
resuming, then even mcast can be resumed after a interruption. This is the next
thing TODO.