Client: Load snm context that was previously saved
The client provides a callback function to ll_load_append_ex, which allows it
to extract all lines begining with '#' to be processed by it.

In turn that logic extracts MaxDataSeqNumGot snm context, from the context file
being loaded.

In future, other members of snm context could be saved by save_context and
inturn loaded by load_context.

This will now allow the client logic to be interrupted and resumed, even when
it is in the mcast phase. Additional logics to support this was commited
earlier today.