Client: Fix a oversight wrt snm contextisation of logics previously
Other than the main logic and signal handler no one else should be using the
snm context structure directly. All others should use the snm context based on
the pointer to same passed to them as part of their arguments.

When the different logics on the client side was snm contextised 1 or 2 days
back, as part of that even opening of data file for writing was also moved into
a snm contextised function. However at one place in this logic, instead of
using the passed snm context, it was using the global snm context directly.
This has been fixed now.