Client: DoneModes & optionally autoAdjust RunModes
Have added a new context variable called DoneModes. The corresponding bit in
this is set at the end of each phase, irrespective of whether it was actively
run or skipped.

The same is inturn saved and restored as part of the context file logic.

Additionally there is a compile time option to enable a logic which will use
the saved DoneModes value to setup the RunModes when a context is loaded into
the program. This ensures that if the client is quit in the middle of transfer,
then when the program is resumed using the corresponding saved context file,
the program will automatically go back to the phase in which the program was
running when it was forced to quit.

By default this autoadjusting of runmodes from donemodes, when a context is
being restored/resumed, is enabled.