Client: remove old llLostPkts, use one in snm context
Fixed a oversight in main, which is harmless wrt functionality, but which can
lead to wrong prints on the console.

This oversight was introduced when the llLostPkts was snm contextised in
general 2 odd days back. The ll print and free logic continued to use the
locally defined one in main. Now they use the llLostPkts in snm.