Client Cleanup - recent args usage, use _account func
Add the usage info for the recently added args --nwgroup and --runmodes.

When ever there is packet losses identified as part of the normal mcast recv
flow, now it uses the _account_lost_packets func which does the necessary
updates, instead of doing those individual operations on its own.

This _account... function was created so that it could be used from multiple
locations like

* the normal mcast recv flow packet loss identification,
* identification of lost packets at the end, based on info in mcast stop
  command

TODO: This should also be used, if a client is interruped during mcast transfer
and then resumed into unicast phase.