TODO wrt possible unidentified packet loss situation,
if user forces runmodes wrongly i.e if a client interrupted in mcast transfer
phase is forced to resume in unicast phases, then all packets following the
interruption till the end of transfer will be lost for ever, with no way for
client to identify this situation.

Have to fix this.