check-image: lostRange, betterDeltaErrorIdent, TotalErrors
Instead of printing each individual lost packet/block info, now it prints lost
blocks as ranges.

More Intelligent potential identification of wrong/missing blocks, by using
Delta between cur and expected prev block's id in the blocks' data. Now it
flags a warning wrt the current block, if the difference is other than 1 as
long as it doesnt match the case for the current block being all 0 (i.e not
recieved at all), because a not recieved block is already captured by the
cur!=ref logic.

NOTE: Currently if the block's seq number location contains 0, it is assumed to
be a not recieved block. Which ideally should be fine, if there is no other bug
in the transfer or write logic wrt how this file being checked was created in
the first place.

Now a counter is maintained as to how many blocks in the file had issues.