URReq has TotalBlocksInvolved, Update MaxDataSeq only if reqd
* Account total blocks involved properly

If the user interrupts a transfer in the middle of mcast transfer and then
forces the program to directly jump into unicast recover phase on resume, then
previously the logic wouldn't have accounted for all packets from the time of
interruption, till end of the transfer. This has been fixed now as mentioned
below.

The server now sends the TotalBlocksInvolved as a additional parameter.

The snm context now maintains a additional variable which tracks the max data
seq number the logic has processed till that point in time. It could either be
a valid data packet it has processed or it could be the last packet involved in
the overall transfer, which one comes to know through mcast_stop or urreq
commands/packets. In turn if required the lost packet ranges is updated.

The MaxDataSeqNumProcessed is saved and restored as part of save and load
context logics. <<== TODO.

With the above logic and context, even if the user forces a interrupted program
to jump from mcast to ucast directly when the program is resumed, the logic can
know about the lost packets after the program was originally interrupted and
inturn add it to the list of lost packet ranges, when it resumes and recieves a
URReq packet.

* Update MaxDataSeqNum related variables only if required.

Previously each time a data packet was recieved, the MaxDataSeqNumGot was
getting updated with the seq number in that data packet. However if it turns
out to be a old out of seq packet, then the value in MaxDataSeqNumGot would
have been wrong. This has been fixed now.

Similar logic is also used for MaxDataSeqNumProcessed.