Server: Status Cleanup, flush, _print
Now each progress update function calls a predefine _before and _after
functions.

_after inturn flushes the statusFile.

Added a _print function, which prints the passed message to the statusFile,
provided the msg lvl is below a predefined threshold. The default lvl is setup
to print a message, if the message didn't specify its lvl explicitly.

Progress update functions now call this internal _print to print the progress
updates.

_before and _after have marker messages, which are disabled as of now, by
setting a msg lvl which is larger than the predefined threshold.