Server:send-mcast: save context in sigint, intelligent loading
A signal handler has been added for SIGINT. So if the program is forced to quit
using SIGINT, then it will allow the current mcast transfer context to be
saved, before the program is quit.

Previously I was quiting the program if the specified context was empty or
didn't contain any relevant mcast context data.

However if I want to specify as to where the context should be saved, then I
shouldn't quit in this case, instead of it, I should ignore the lack of context
during loading, and allow the logic to continue, so that if the program is
forced to quit later, it can save the context into this specified context file.

This also means that if I want to control the location into which the context
will be saved, and at same time if I want the program to work from the
beginning wrt mcast transfer, then I should pass a empty or non-existant
context file. Else the program will start from the location specified in such a
context file.

Notes updated wrt the next release's release note info for this feature.