Server:PI:Status FixCntBug & Name+ProperLostPkts
The status logging func ucast_pi was wrongly expecting the count of PIReq Pkts
recieved from client to be saved with the key 'Cnt' in clientsDB. But actually
it is saved using the key 'cnt', so the nw-recover program would have failed at
the end of PI phase, when trying to write the summary progress update into
status file. This oversight has been fixed now.

Now LostPkts is stored using the key 'lostpkts' instead of 'LostPkts', so that
key naming is consistant, rather that is what led to the 'Cnt' oversight in the
first place.

Also now the actual ucast pi logic extracts the Name and LostPkts info from the
PIReq packet properly (previously the full PIReq packet content was getting
assigned into LostPkts key of clientsDB) and saves the same into clientsDB.
Inturn the same is saved in the summary status file.