Server:send-mcast: seek if not beginning from start
With the addition of --startblock and restore context logics, now the
send-mcast program may require to start transfering from somewhere inbetween in
the file rather than from the beginning. TO account for this one has to seek
the file to the required block at the beginning. This has been handled now.