Server:Status: SET toIdentify GroupInfo
Identify status logging in one shot wrt some group/set of info. Such group info
loggin is now prefixed with a line starting with SET and giving some summary
detail about the group and the count of entries.

Currently this is used to identify PI status at end of PI Phase as well as to
identify info about the remaining group/set of clients at the end of each UR
phase attempt.