Client: Rejoin if mcast channel silent for long
Now the mcast_recv takes the snm context has a additional optional last
argument.

If the mcast_recv hasn't recieved any mcast packets for a predefined long time,
then the mcast socket is made to join the mcast channel once again.

NOTE:

When mcast socket is created the first time, if mcast join fails, the program
will quit.

However when rejoining mcast group, due to silent mcast channel, Even if the
mcast rejoining fails, it is ignored after logging the failure. This is to
ensure that if the interface being used for joining the mcast group is down, it
doesn't force the program to quit.